refactor : change dir of chatService
