+++ src/main/java/com/chatty/practice/suhyeon/tutorial/chatapp/controller/ChatController.java
@@ -1,11 +1,11 @@
 package com.chatty.practice.suhyeon.tutorial.chatapp.controller;
 
 import com.chatty.practice.suhyeon.tutorial.chatapp.dto.ChatMessage;
+import com.chatty.practice.suhyeon.tutorial.chatapp.service.ChatService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
